some repositories added

+++ SchoolSystem/SchoolSystem.Repository/Services/IStudentRepository.cs
@@ -4,11 +4,11 @@ namespace SchoolSystem.Repository.Services;
 
 public interface IStudentRepository
 {
-    Task<Student> SelectStudentByIdAsync(int id);
+    Task<Student> SelectStudentByIdAsync(long id);
     Task<List<Student>> SelectAllStudentsAsync(bool includeTeacher, bool includeClass, int skip, int take);
     Task InsertStudentAsync(Student student);
     Task UpdateStudentAsync(Student student);
-    Task DeleteStudentAsync(int id);
+    Task DeleteStudentAsync(long id);
 }
 
 /***
